src/third_party/blink/web_tests/platform/win7/external/wpt/event-timing/dblclick-expected.txt

This is a testharness.js-based test.
FAIL Event Timing dblclick. assert_greater_than: The entry's processingStart should be greater than startTime. expected a number greater than 470.9100000000035 but got 470.9100000000035
Harness: the test ran to completion.
